requirements.txt
numpy
matplotlib
torch
torchvision
scikit-learn
pandas
tqdm
transformers
pyyaml